refactor: remove internal/pkg/core/backoff wrapper, use pkg/backoff directly
All callers now import pkg/backoff instead of going through the
internal/pkg/core/backoff indirection layer.
