--- internal/pkg/core/backoff/backoff.go
@@ -1,9 +1,0 @@
-// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-// or more contributor license agreements. Licensed under the Elastic License 2.0;
-// you may not use this file except in compliance with the Elastic License 2.0.
-
-package backoff
-
-import publicbackoff "github.com/elastic/elastic-agent/pkg/backoff"
-
-type Backoff = publicbackoff.Backoff
