+++ internal/pkg/fleetapi/uploader/uploader.go
@@ -17,7 +17,7 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/elastic/elastic-agent/internal/pkg/core/backoff"
+	"github.com/elastic/elastic-agent/pkg/backoff"
 	"github.com/elastic/elastic-agent/internal/pkg/core/monitoring/config"
 	"github.com/elastic/elastic-agent/internal/pkg/fleetapi/client"
 )
